requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosqlite==0.19.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
async-timeout==5.0.1
asyncio==3.4.3
asyncpg==0.30.0
bcrypt==4.0.1
black==24.10.0
bleak==0.22.3
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
contourpy==1.3.1
coverage==7.6.10
cryptography==44.0.0
cycler==0.12.1
dill==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.109.2
fonttools==4.55.5
greenlet==3.1.1
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
kiwisolver==1.4.8
lru-dict==1.2.0
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.2.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
psutil==6.1.1
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyJWT==2.8.0
pylint==3.3.3
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreBluetooth==10.3.2
pyobjc-framework-libdispatch==10.3.2
pyparsing==3.2.1
pypng==0.20220715.0
pytest==8.0.0
pytest-asyncio==0.23.8
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
qrcode==7.4.2
redis==5.0.1
rsa==4.9
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.36.3
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.27.1
requests
python-jose[cryptography]
passlib[bcrypt]==1.7.4
jinja2==3.1.3
email-validator==2.1.0.post1

aiosqlite>=0.17.0  # For async SQLite
asyncpg>=0.26.0